[MeasureGui] Move measure cmd and task into measure module
...to avoid core depending on modules
